Change to size_t for stmt ref

+++ Team32/Code32/src/spa/src/Common/TypeDefs.h
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-typedef int StmtRef;
+typedef size_t StmtRef;
 typedef unsigned long long int ConstVal;
 typedef string VarRef;
 typedef string ProcRef;
